test-requirements.txt
-r ./requirements.txt
codecov
flake8
pytest >= 2.8
pytest-cov
pytest-tornado
